Remove dependency on tokens in sentence fragmenter version 2
Updated util functions and sentence fragmenter to operate on string_view using a sliding window rather than tokens and return a boolean rather than a status, as UTF-8 is assumed now. New version is in sentence_fragmenter_v2, with additional tests for utils and fragment boundary match in sentence_fragmenter_v2_test, including tests that check the state transitions of the Advance function.

The new op takes in only “doc,” a tensor of the strings. Added version 2 of the sentence breaking op to the python API, wrapped in a HeuristicBasedSentenceBreaker class, and updated tests for sentence_breaking_op_v2.

PiperOrigin-RevId: 316925297