Update StateBasedSentenceBreaker handling of text input tensors.
Updated StateBasedSentenceBreaker break_sentences_with_offsets to appropriately handle input when the doc is a tensor, previously input was converted to a ragged tensor and the flat values were passed into the sentence fragmenter. Now, a tensor input is passed into the sentence fragmenter directly and we only get the flat values of a ragged tensor input.

PiperOrigin-RevId: 322387130