Add a subwords tokenizer tutorial to text/examples.
This tutorial generates a subword vocabularies and builds an end-to-end text encoder and decoder, and then exports that as a saved model.

The resulting saved_model is available at:

https://storage.googleapis.com/download.tensorflow.org/models/ted_hrlr_translate_pt_en_converter.zip

So it can be used by the text generation and translation tutorials, and they can both point here for the "How-to".

fixes: tensorflow/datasets#2879
fixes: tensorflow/datasets#2891
PiperOrigin-RevId: 355194834