`regex_split_with_offsets` & `BertTokenizer`: add support for scalar inputs.
For `regex_split_with_offsets`: when receiving a scalar input, the output will be vectors. This should honor the expected invariant that for an input with rank N the function returns an output with rank N+1.

For `BertTokenizer`: when receiving a scalar input, the output will be a ragged matrix. Previously, given the lack of support for scalars in `regex_split_with_offsets`, the output would be a ragged 3-D tensor. This should honor the expected invariant that for an input with rank N the tokenizer returns an output with rank N+2.

PiperOrigin-RevId: 427601172