Update word_embeddings.ipynb with Time Series as default dashboard.
With the TensorBoard 2.10 release the "Time Series" dashboard will be the default dashboard for logs with scalars, histograms, or images events. We are updating external-facing documentation to reflect this change.

Mirroring work in the TensorBoard repository (https://github.com/tensorflow/tensorboard/pull/5861) this change updates notebook documentation to use the "Time Series" dashboard where previously it used the "Scalars" dashboard.

Notably:
* Updated images to use "Time Series" and set them to 1600x800 pixels to be consistent with other changes.
PiperOrigin-RevId: 511784849