Update TF versions and scripts to allow consistently building against tf-nightly.
Currently the master branch contains a patch and other changes
that is needed to build against a later version of TF than the
2.13 release (and even the 2.14 rc0).  Bumped up the default
commit in the WORKSPACE file to allow a build, and bumped up
the release version to 2.13 for now (was previously 2.12).

Sometimes the tf-nightly build reports a `tf.__git_version__` of
"unknown" (occasionally on intel, frequently on ARM), causing the
build scripts to fail and abort.  Modified the `prepare_tf_dep.sh`
to account for this, defaulting to the latest tf-nightly source in
such a case.

Fixes #1202, #1195.

PiperOrigin-RevId: 559233072